+++ src/core/geometry.cpp
@@ -3,25 +3,25 @@
 #include <algorithm>
 #include <stdexcept>
 
-Vertex::Vertex(long x, long y) : x(x), y(y) {}
+Vertex::Vertex(int x, int y) : x(x), y(y) {}
 
-long Vertex::getX() const {
+int Vertex::getX() const {
     return x;
 }
 
-long Vertex::getY() const {
+int Vertex::getY() const {
     return y;
 }
 
-void Vertex::setX(long x) {
+void Vertex::setX(int x) {
     this->x = x;
 }
 
-void Vertex::setY(long y) {
+void Vertex::setY(int y) {
     this->y = y;
 }
 
-void Vertex::translate(long dx, long dy) {
+void Vertex::translate(int dx, int dy) {
     this->x += dx;
     this->y += dy;
 }
@@ -63,17 +63,17 @@ Vertex Polygon::getVertex(int index) const {
     return vertices[index];
 }
 
-void Polygon::translate(long dx, long dy) {
+void Polygon::translate(int dx, int dy) {
     for (auto& vertex : vertices) {
         vertex.translate(dx, dy);
     }
 }
 
-long Polygon::getMaxX() const {
+int Polygon::getMaxX() const {
     if (vertices.empty()) {
         throw std::runtime_error("Polygon has no vertices");
     }
-    long maxX = vertices[0].getX();
+    int maxX = vertices[0].getX();
     for (const auto& vertex : vertices) {
         if (vertex.getX() > maxX) {
             maxX = vertex.getX();
@@ -82,11 +82,11 @@ long Polygon::getMaxX() const {
     return maxX;
 }
 
-long Polygon::getMinX() const {
+int Polygon::getMinX() const {
     if (vertices.empty()) {
         throw std::runtime_error("Polygon has no vertices");
     }
-    long minX = vertices[0].getX();
+    int minX = vertices[0].getX();
     for (const auto& vertex : vertices) {
         if (vertex.getX() < minX) {
             minX = vertex.getX();
@@ -95,11 +95,11 @@ long Polygon::getMinX() const {
     return minX;
 }
 
-long Polygon::getMaxY() const {
+int Polygon::getMaxY() const {
     if (vertices.empty()) {
         throw std::runtime_error("Polygon has no vertices");
     }
-    long maxY = vertices[0].getY();
+    int maxY = vertices[0].getY();
     for (const auto& vertex : vertices) {
         if (vertex.getY() > maxY) {
             maxY = vertex.getY();
@@ -108,11 +108,11 @@ long Polygon::getMaxY() const {
     return maxY;
 }
 
-long Polygon::getMinY() const {
+int Polygon::getMinY() const {
     if (vertices.empty()) {
         throw std::runtime_error("Polygon has no vertices");
     }
-    long minY = vertices[0].getY();
+    int minY = vertices[0].getY();
     for (const auto& vertex : vertices) {
         if (vertex.getY() < minY) {
             minY = vertex.getY();
@@ -133,7 +133,7 @@ bool Polygon::operator==(const Polygon& other) const {
     return true;
 }
 
-bool Polygon::containsPoint(long x, long y) const {
+bool Polygon::containsPoint(int x, int y) const {
     int n = vertices.size();
     if (n < 3) {
         return false; // A polygon must have at least 3 vertices
@@ -152,14 +152,14 @@ bool Polygon::containsPoint(long x, long y) const {
     return inside;
 }
 
-set<pair<long, long>> Polygon::getContainedPixels() const {
-    long minX = getMinX();
-    long maxX = getMaxX();
-    long minY = getMinY();
-    long maxY = getMaxY();
-    set<pair<long, long>> pixels;
-    for (long x = minX; x <= maxX; ++x) {
-        for (long y = minY; y <= maxY; ++y) {
+set<pair<int, int>> Polygon::getContainedPixels() const {
+    int minX = getMinX();
+    int maxX = getMaxX();
+    int minY = getMinY();
+    int maxY = getMaxY();
+    set<pair<int, int>> pixels;
+    for (int x = minX; x <= maxX; ++x) {
+        for (int y = minY; y <= maxY; ++y) {
             if (containsPoint(x, y)) {
                 pixels.insert({x, y});
             }
@@ -168,8 +168,8 @@ set<pair<long, long>> Polygon::getContainedPixels() const {
     return pixels;
 }
 
-set<pair<long, long>> Polygon::getBorderPixels() const {
-    set<pair<long, long>> borderPixels;
+set<pair<int, int>> Polygon::getBorderPixels() const {
+    set<pair<int, int>> borderPixels;
 
     int n = vertices.size();
 
@@ -178,21 +178,21 @@ set<pair<long, long>> Polygon::getBorderPixels() const {
         const Vertex& vj = vertices[j];
 
         // Simple line drawing algorithm (Bresenham's line algorithm)
-        long x0 = vi.getX();
-        long y0 = vi.getY();
-        long x1 = vj.getX();
-        long y1 = vj.getY();
+        int x0 = vi.getX();
+        int y0 = vi.getY();
+        int x1 = vj.getX();
+        int y1 = vj.getY();
 
-        long dx = abs(x1 - x0);
-        long dy = abs(y1 - y0);
-        long step_x = (x0 < x1) ? 1 : -1;
-        long step_y = (y0 < y1) ? 1 : -1;
-        long err = dx - dy;
+        int dx = abs(x1 - x0);
+        int dy = abs(y1 - y0);
+        int step_x = (x0 < x1) ? 1 : -1;
+        int step_y = (y0 < y1) ? 1 : -1;
+        int err = dx - dy;
 
         while (true) {
             borderPixels.insert({x0, y0});
             if (x0 == x1 && y0 == y1) break;
-            long e2 = 2 * err;
+            int e2 = 2 * err;
             if (e2 > -dy) {
                 err -= dy;
                 x0 += step_x;
